requirements.txt
asgiref==3.7.2
boto3==1.26.146
botocore==1.29.146
Django==4.2.1
psycopg2-binary 
djangorestframework==3.14.0
jmespath==1.0.1
python-dateutil==2.8.2
pytz==2023.3
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==1.26.16
gunicorn
PyJWT